rpcutils: move RFQ gRPC marshal functionality from taprpc
Move RFQ related gRPC marshalling functionality from the taprpc package
to the rpcutils package as preparation for future modularization.
