+++ rpcutils/rfq_marshal.go
@@ -1,14 +1,17 @@
-package rfqrpc
+package rpcutils
 
 import (
 	"fmt"
 	"math/big"
 
 	"github.com/lightninglabs/taproot-assets/rfqmath"
+	"github.com/lightninglabs/taproot-assets/taprpc/rfqrpc"
 )
 
-// UnmarshalFixedPoint converts an RPC FixedPoint to a BigIntFixedPoint.
-func UnmarshalFixedPoint(fp *FixedPoint) (*rfqmath.BigIntFixedPoint, error) {
+// UnmarshalRfqFixedPoint converts an RPC FixedPoint to a BigIntFixedPoint.
+func UnmarshalRfqFixedPoint(fp *rfqrpc.FixedPoint) (*rfqmath.BigIntFixedPoint,
+	error) {
+
 	// Return an error is the scale component of the fixed point is greater
 	// than the max value of uint8.
 	if fp.Scale > 255 {
